DRY card images
