+++ src/GameScreen/Card.js
@@ -1,89 +1,11 @@
 /* eslint-disable jsx-a11y/no-static-element-interactions */
 /* eslint-disable jsx-a11y/click-events-have-key-events */
-/* eslint-disable import/no-duplicates */
 
 import React, { useContext } from 'react';
 import PropTypes from 'prop-types';
 import BoardContext from './BoardContext';
 import '../css/Cards.css';
-
-import apple from '../images/cards/apple.png';
-import bajun from '../images/cards/catbajun.png';
-import bat from '../images/cards/bat.png';
-import backgroundCommon from '../images/cards/background_common1.jpg';
-import backgroundAttack from '../images/cards/background_common2.jpg';
-import bereginya from '../images/cards/bitva-cardbase.jpg';
-import bogatyr from '../images/cards/bitva-cardbase.jpg';
-import bolotnitsa from '../images/cards/bolotnitsa.png';
-import bowArrow from '../images/cards/bitva-cardbase.jpg';
-import bulat from '../images/cards/sword.png';
-import chemise from '../images/cards/bitva-cardbase.jpg';
-import chickenLegsHut from '../images/cards/izba.png';
-import clairvoyance from '../images/cards/bitva-cardbase.jpg';
-import crown from '../images/cards/bitva-cardbase.jpg';
-import dolly from '../images/cards/bitva-cardbase.jpg';
-import earthquake from '../images/cards/bitva-cardbase.jpg';
-import forestMushroom from '../images/cards/mushrooms.png';
-import horsemanBlack from '../images/cards/blackrider.png';
-import horsemanRed from '../images/cards/redrider.png';
-import horsemanWhite from '../images/cards/whiterider.png';
-import kikimora from '../images/cards/kikimora.png';
-import lizard from '../images/cards/lizard.png';
-import magicTree from '../images/cards/tree.png';
-import magicMirror from '../images/cards/bitva-cardbase.jpg';
-import malachiteBox from '../images/cards/malachitebox.png';
-import mortar from '../images/cards/bitva-cardbase.jpg';
-import plateMail from '../images/cards/bitva-cardbase.jpg';
-import raven from '../images/cards/raven.png';
-import russianOven from '../images/cards/bitva-cardbase.jpg';
-import shieldLarge from '../images/cards/largeshield.png';
-import shieldSmall from '../images/cards/smallshield.png';
-import sivka from '../images/cards/warhorse.png';
-import skullLantern from '../images/cards/bitva-cardbase.jpg';
-import turningPotion from '../images/cards/potion.png';
-import warhorse from '../images/cards/warhorse.png';
-import waterDead from '../images/cards/deadwater.png';
-import waterLiving from '../images/cards/livingwater.png';
-import wolf from '../images/cards/wolf.png';
-
-const imagesCards = {
-    apple,
-    bajun,
-    bat,
-    bereginya,
-    bogatyr,
-    bolotnitsa,
-    bowArrow,
-    bulat,
-    chemise,
-    chickenLegsHut,
-    clairvoyance,
-    crown,
-    dolly,
-    earthquake,
-    forestMushroom,
-    horsemanBlack,
-    horsemanRed,
-    horsemanWhite,
-    kikimora,
-    lizard,
-    magicTree,
-    magicMirror,
-    malachiteBox,
-    mortar,
-    plateMail,
-    raven,
-    russianOven,
-    shieldLarge,
-    shieldSmall,
-    sivka,
-    skullLantern,
-    turningPotion,
-    warhorse,
-    waterDead,
-    waterLiving,
-    wolf,
-};
+import imagesCards from '../cardImages';
 
 function cardClass(type, background, dragging) {
     if (!dragging && type === 'item') {
@@ -96,8 +18,8 @@ function cardClass(type, background, dragging) {
 }
 export function backgroundImg(category) {
     if (category === 'attackItems' || category === 'attack') {
-        return backgroundAttack;
-    } return backgroundCommon;
+        return imagesCards.backgroundAttack;
+    } return imagesCards.backgroundCommon;
 }
 
 export function cardOrigin(dragging, card) {
